SME-35 refactoring - LogoutwithoutRedirect moved to userprofile/actions, default login redirect to destination added, topbar moving logo while transition fixed
